requirements.txt
numpy
scikit-learn
sentence-transformers
requests
python-dateutil
tqdm
streamlit
python-dotenv
langchain
langchain_ollama
streamlit
pandas
